tools/ludo/requirements.txt
httpx>=0.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
typer>=0.12.0
rich>=13.7.0
